Create different folders for the testCode and Code.
