+++ llvm/.gitignore
@@ -1,0 +1,2 @@
+LibDefExtractor/bin/
+LibDefExtractor/obj/
